Address PR comments and issues
