+++ docs/resources/integration_credential.md
@@ -1,10 +1,10 @@
 ---
-page_title: "genesyscloud_credential Resource - terraform-provider-genesyscloud"
+page_title: "genesyscloud_integration_credential Resource - terraform-provider-genesyscloud"
 subcategory: ""
 description: |-
   Genesys Cloud Credential
 ---
-# genesyscloud_credential (Resource)
+# genesyscloud_integration_credential (Resource)
 
 Genesys Cloud Credential
 
@@ -20,7 +20,7 @@ The following Genesys Cloud APIs are used by this resource. Ensure your OAuth Cl
 ## Example Usage
 
 ```terraform
-resource "genesyscloud_credential" "credential1" {
+resource "genesyscloud_integration_credential" "credential1" {
   name                 = "test-credential"
   credential_type_name = "basicAuth" //Example type
   fields = {
@@ -36,11 +36,11 @@ resource "genesyscloud_credential" "credential1" {
 
 ### Required
 
-- **credential_type_name** (String) Credential type name.
+- **credential_type_name** (String) Credential type name. Use [GET /api/v2/integrations/credentials/types](https://developer.genesys.cloud/api/rest/v2/integrations/#get-api-v2-integrations-credentials-types) to see the list of available integration credential types.
 
 ### Optional
 
-- **fields** (Map of String) Credential fields. Different credential types require different fields. Missing any correct required fields will result API request failure. Check out the specific credential type schema to find out what fields are required.
+- **fields** (Map of String, Sensitive) Credential fields. Different credential types require different fields. Missing any correct required fields will result API request failure. Use [GET /api/v2/integrations/credentials/types](https://developer.genesys.cloud/api/rest/v2/integrations/#get-api-v2-integrations-credentials-types) to check out the specific credential type schema to find out what fields are required.
 - **id** (String) The ID of this resource.
 - **name** (String) Credential name.
