+++ examples/resources/genesyscloud_integration_credential/apis.md
