+++ genesyscloud/provider.go
@@ -66,7 +66,6 @@ func New(version string) func() *schema.Provider {
 				"genesyscloud_architect_datatable_row": resourceArchitectDatatableRow(),
 				"genesyscloud_auth_role":               resourceAuthRole(),
 				"genesyscloud_auth_division":           resourceAuthDivision(),
-				"genesyscloud_credential":              resourceCredential(),
 				"genesyscloud_group":                   resourceGroup(),
 				"genesyscloud_group_roles":             resourceGroupRoles(),
 				"genesyscloud_idp_adfs":                resourceIdpAdfs(),
@@ -77,6 +76,7 @@ func New(version string) func() *schema.Provider {
 				"genesyscloud_idp_ping":                resourceIdpPing(),
 				"genesyscloud_idp_salesforce":          resourceIdpSalesforce(),
 				"genesyscloud_integration":             resourceIntegration(),
+				"genesyscloud_integration_credential":  resourceCredential(),
 				"genesyscloud_location":                resourceLocation(),
 				"genesyscloud_routing_email_domain":    resourceRoutingEmailDomain(),
 				"genesyscloud_routing_email_route":     resourceRoutingEmailRoute(),
