+++ genesyscloud/resource_exporter.go
@@ -118,7 +118,6 @@ func getResourceExporters(filter []string) map[string]*ResourceExporter {
 		"genesyscloud_architect_datatable_row": architectDatatableRowExporter(),
 		"genesyscloud_auth_division":           authDivisionExporter(),
 		"genesyscloud_auth_role":               authRoleExporter(),
-		"genesyscloud_credential":              credentialExporter(),
 		"genesyscloud_group":                   groupExporter(),
 		"genesyscloud_group_roles":             groupRolesExporter(),
 		"genesyscloud_idp_adfs":                idpAdfsExporter(),
@@ -129,6 +128,7 @@ func getResourceExporters(filter []string) map[string]*ResourceExporter {
 		"genesyscloud_idp_ping":                idpPingExporter(),
 		"genesyscloud_idp_salesforce":          idpSalesforceExporter(),
 		"genesyscloud_integration":             integrationExporter(),
+		"genesyscloud_integration_credential":  credentialExporter(),
 		"genesyscloud_location":                locationExporter(),
 		"genesyscloud_routing_email_domain":    routingEmailDomainExporter(),
 		"genesyscloud_routing_email_route":     routingEmailRouteExporter(),
