+++ genesyscloud/resource_genesyscloud_tf_export_test.go
@@ -151,7 +151,6 @@ func generateTfExportResource(
 			"genesyscloud_architect_datatable_row",
 			"genesyscloud_auth_role",
 			"genesyscloud_auth_division",
-			"genesyscloud_credential",
 			"genesyscloud_group",
 			"genesyscloud_group_roles",
 			"genesyscloud_idp_adfs",
@@ -162,6 +161,7 @@ func generateTfExportResource(
 			"genesyscloud_idp_ping",
 			"genesyscloud_idp_salesforce",
 			"genesyscloud_integration",
+			"genesyscloud_integration_credential",
 			"genesyscloud_location",
 			"genesyscloud_routing_email_domain",
 			"genesyscloud_routing_email_route",
